chore(server): move the definition of `NoRoute` and `Endpoint::to_service` into service/
