+++ src/endpoint/mod.rs
@@ -27,7 +27,7 @@ pub(crate) mod with;
 // re-exports
 pub use self::body::body;
 pub use self::context::{EndpointContext, Segments};
-pub use self::endpoint::{Endpoint, IntoEndpoint, NoRoute};
+pub use self::endpoint::{Endpoint, IntoEndpoint};
 pub use self::header::{header, header_opt, EmptyHeader};
 #[doc(inline)]
 pub use self::method::MatchMethod;
