+++ src/service/mod.rs
@@ -1,7 +1,9 @@
 //! Components to construct Hyper's service.
 
+mod endpoint_ext;
 mod server;
 mod service;
 
+pub use self::endpoint_ext::*;
 pub use self::server::*;
 pub use self::service::*;
